+++ packages/@expo/cli/src/export/embed/exportEmbedAsync.ts
@@ -4,11 +4,9 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { ExpoConfig, getConfig, modifyConfigAsync, PackageJSONConfig } from '@expo/config';
+import { ExpoConfig, getConfig } from '@expo/config';
 import getMetroAssets from '@expo/metro-config/build/transform-worker/getAssets';
-import spawnAsync from '@expo/spawn-async';
 import assert from 'assert';
-import chalk from 'chalk';
 import crypto from 'crypto';
 import fs from 'fs';
 import { sync as globSync } from 'glob';
@@ -16,20 +14,12 @@ import Server from 'metro/src/Server';
 import splitBundleOptions from 'metro/src/lib/splitBundleOptions';
 import output from 'metro/src/shared/output/bundle';
 import type { BundleOptions } from 'metro/src/shared/types';
-import { execSync } from 'node:child_process';
 import path from 'path';
 import resolveFrom from 'resolve-from';
 
 import { deserializeEagerKey, getExportEmbedOptionsKey, Options } from './resolveOptions';
-import {
-  isExecutingFromXcodebuild,
-  logInXcode,
-  logMetroErrorInXcode,
-  warnInXcode,
-} from './xcodeCompilerLogger';
-import { disableNetwork } from '../../api/settings';
+import { isExecutingFromXcodebuild, logMetroErrorInXcode } from './xcodeCompilerLogger';
 import { Log } from '../../log';
-import { isSpawnResultError } from '../../start/platforms/ios/xcrun';
 import { DevServerManager } from '../../start/server/DevServerManager';
 import { MetroBundlerDevServer } from '../../start/server/metro/MetroBundlerDevServer';
 import { loadMetroConfigAsync } from '../../start/server/metro/instantiateMetro';
@@ -43,10 +33,8 @@ import { getMetroDirectBundleOptionsForExpoConfig } from '../../start/server/mid
 import { stripAnsi } from '../../utils/ansi';
 import { copyAsync, removeAsync } from '../../utils/dir';
 import { env } from '../../utils/env';
-import { CommandError } from '../../utils/errors';
 import { setNodeEnv } from '../../utils/nodeEnv';
 import { isEnableHermesManaged } from '../exportHermes';
-import { exportApiRoutesStandaloneAsync } from '../exportStaticAsync';
 import { persistMetroAssetsAsync } from '../persistMetroAssets';
 import { copyPublicFolderAsync } from '../publicFolder';
 import {
@@ -55,6 +43,7 @@ import {
   getFilesFromSerialAssets,
   persistMetroFilesAsync,
 } from '../saveAssets';
+import { exportStandaloneServerAsync } from './exportServer';
 
 const debug = require('debug')('expo:export:embed');
 
@@ -170,385 +159,6 @@ export async function exportEmbedInternalAsync(projectRoot: string, options: Opt
         })
       : null,
   ]);
-}
-
-async function dumpDeploymentLogs(projectRoot: string, logs: string, name = 'deploy') {
-  const outputPath = path.join(projectRoot, `.expo/logs/${name}.log`);
-  await fs.promises.mkdir(path.dirname(outputPath), { recursive: true });
-  debug('Dumping server deployment logs to: ' + outputPath);
-  await fs.promises.writeFile(outputPath, logs);
-  return outputPath;
-}
-
-function getCommandBin(command: string) {
-  try {
-    return execSync(`command -v ${command}`, { stdio: 'pipe' }).toString().trim();
-  } catch {
-    return null;
-  }
-}
-
-async function runServerDeployCommandAsync(
-  projectRoot: string,
-  {
-    distDirectory,
-    deployScript,
-  }: { distDirectory: string; deployScript: { scriptName: string; script: string } | null }
-): Promise<string | false> {
-  const logOfflineError = () => {
-    const manualScript = deployScript
-      ? `npm run ${deployScript.scriptName}`
-      : `npx eas deploy --export-dir ${distDirectory}`;
-
-    logMetroErrorInXcode(
-      projectRoot,
-      chalk.red`Running CLI in offline mode, skipping server deployment. Deploy manually with: ${manualScript}`
-    );
-  };
-  if (env.EXPO_OFFLINE) {
-    logOfflineError();
-    return false;
-  }
-  // TODO: Test error cases thoroughly since they run at the end of a build:
-  // - EAS not installed.
-  // - EAS not configured.
-  // - Build from Xcode locally.
-  // - Build from Xcode with EAS.
-  // - Network error.
-  // - Custom deploy script `deploy.native`.
-
-  const globalBin = getCommandBin('eas');
-
-  if (!globalBin) {
-    // This should never happen from EAS Builds.
-    // Possible to happen when building locally with `npx expo run`
-    logMetroErrorInXcode(
-      projectRoot,
-      `eas-cli is not installed globally, skipping server deployment. Install EAS CLI with 'npm install -g eas-cli'.`
-    );
-    return false;
-  }
-  debug('Found eas-cli:', globalBin);
-
-  let json: any;
-  try {
-    let results: spawnAsync.SpawnResult;
-
-    const spawnOptions: spawnAsync.SpawnOptions = {
-      cwd: projectRoot,
-      // Ensures that errors can be caught.
-      stdio: 'pipe',
-    };
-    // TODO: Support absolute paths in EAS CLI
-    const exportDir = path.relative(projectRoot, distDirectory);
-    if (deployScript) {
-      logInXcode(`Using custom server deploy script: ${deployScript.scriptName}`);
-      // Amend the path to try and make the custom scripts work.
-
-      results = await spawnAsync(
-        'npm',
-        ['run', deployScript.scriptName, `--export-dir=${exportDir}`],
-        spawnOptions
-      );
-    } else {
-      logInXcode('Deploying server to link with client');
-
-      // results = DEPLOYMENT_SUCCESS_FIXTURE;
-      results = await spawnAsync(
-        'node',
-        [globalBin, 'deploy', '--non-interactive', '--json', `--export-dir=${exportDir}`],
-        spawnOptions
-      );
-
-      // TODO: Print dashboard URL.
-      debug('Server deployment stdout:', results.stdout);
-
-      // Send stderr to stderr. stdout is parsed as JSON.
-      if (results.stderr) {
-        process.stderr.write(results.stderr);
-      }
-    }
-
-    const logPath = await dumpDeploymentLogs(projectRoot, results.output.join('\n'));
-
-    try {
-      // {
-      //   "dashboardUrl": "https://staging.expo.dev/projects/80ca6300-4db2-459e-8fde-47bad9c532ff/hosting/deployments",
-      //   "deployment": {
-      //     "identifier": "29bkrcd7ky",
-      //     "url": "https://sep23--29bkrcd7ky.staging.expo.app"
-      //   }
-      // }
-      json = JSON.parse(results.stdout.trim());
-    } catch {
-      logMetroErrorInXcode(
-        projectRoot,
-        `Failed to parse server deployment JSON output. Check the logs for more information: ${logPath}`
-      );
-      return false;
-    }
-  } catch (error) {
-    if (isSpawnResultError(error)) {
-      const output = error.output.join('\n').trim() || error.toString();
-      Log.log(
-        chalk.dim(
-          'An error occurred while deploying server. Logs stored at: ' +
-            (await dumpDeploymentLogs(projectRoot, output, 'deploy-error'))
-        )
-      );
-
-      // Likely a server offline or network error.
-      if (output.match(/ENOTFOUND/)) {
-        logOfflineError();
-        // Print the raw error message to help provide more context.
-        Log.log(chalk.dim(output));
-        // Prevent any other network requests (unlikely for this command).
-        disableNetwork();
-        return false;
-      }
-
-      logInXcode(output);
-      if (output.match(/spawn eas ENOENT/)) {
-        // EAS not installed.
-        logMetroErrorInXcode(
-          projectRoot,
-          `Server deployment failed because eas-cli cannot be accessed from the build script's environment (ENOENT). Install EAS CLI with 'npm install -g eas-cli'.`
-        );
-        return false;
-      }
-
-      if (error.stderr.match(/Must configure EAS project by running/)) {
-        // EAS not configured, this can happen when building a project locally before building in EAS.
-        // User must run `eas init`, `eas deploy`, or `eas build` first.
-
-        // TODO: Should we fail the build here or just warn users?
-        logMetroErrorInXcode(
-          projectRoot,
-          `Skipping server deployment because EAS is not configured. Run 'eas init' before trying again, or disable server output in the project.`
-        );
-        return false;
-      }
-    }
-
-    // Throw unhandled server deployment errors.
-    throw error;
-  }
-
-  // Assert json format
-  assertDeploymentJsonOutput(json);
-
-  // Warn about the URL not being valid. This should never happen, but might be possible with third-parties.
-  if (!URL.canParse(json.deployment.url)) {
-    warnInXcode(`The server deployment URL is not a valid URL: ${json.deployment.url}`);
-  }
-
-  logInXcode(`Server deployed to: ${json.deployment.url}`);
-
-  return json.deployment.url;
-}
-
-type ServerDeploymentResults = {
-  deployment: {
-    url: string;
-  };
-};
-
-function assertDeploymentJsonOutput(json: any): asserts json is ServerDeploymentResults {
-  if (
-    !json ||
-    typeof json !== 'object' ||
-    typeof json.deployment !== 'object' ||
-    typeof json.deployment.url !== 'string'
-  ) {
-    throw new Error(
-      'JSON output of server deployment command are not in the expected format: { deployment: { url: "https://..." } }'
-    );
-  }
-}
-
-function getServerDeploymentScript(
-  scripts: Record<string, string> | undefined,
-  platform: string
-): { scriptName: string; script: string } | null {
-  // Users can overwrite the default deployment script with:
-  // { scripts: { "native:deploy": "eas deploy --json --non-interactive" } }
-  // { scripts: { "native:deploy:ios": "eas deploy" } }
-  // A quick search on GitHub showed that `native:deploy` is not used in any public repos yet.
-  // https://github.com/search?q=%22native%3Adeploy%22+path%3Apackage.json&type=code
-  const DEFAULT_SCRIPT_NAME = 'native:deploy';
-
-  const scriptNames = [DEFAULT_SCRIPT_NAME + ':' + platform, DEFAULT_SCRIPT_NAME];
-
-  for (const scriptName of scriptNames) {
-    if (scripts?.[scriptName]) {
-      return { scriptName, script: scripts[scriptName] };
-    }
-  }
-
-  return null;
-}
-
-async function exportStandaloneServerAsync(
-  projectRoot: string,
-  devServer: MetroBundlerDevServer,
-  {
-    exp,
-    pkg,
-    files,
-    options,
-  }: { exp: ExpoConfig; pkg: PackageJSONConfig; files: ExportAssetMap; options: Options }
-) {
-  if (!options.eager) {
-    await tryRemovingGeneratedOriginAsync(projectRoot, exp);
-  }
-
-  logInXcode('Exporting server');
-
-  // Store the server output in the project's .expo directory.
-  const serverOutput = path.join(projectRoot, '.expo/server', options.platform);
-
-  // Remove the previous server output to prevent stale files.
-  await removeAsync(serverOutput);
-
-  // Export the API routes for server rendering the React Server Components.
-  await exportApiRoutesStandaloneAsync(devServer, {
-    files,
-    platform: 'web',
-  });
-
-  const publicPath = path.resolve(projectRoot, env.EXPO_PUBLIC_FOLDER);
-
-  // Copy over public folder items
-  await copyPublicFolderAsync(publicPath, serverOutput);
-
-  // Copy over the server output on top of the public folder.
-  await persistMetroFilesAsync(files, serverOutput);
-
-  // TODO: This is required for eas cli to detect that the project is dynamic. We need to fix upstream.
-  await fs.promises.mkdir(path.join(serverOutput, 'client'), { recursive: true });
-
-  [...files.entries()].forEach(([key, value]) => {
-    if (value.targetDomain === 'server') {
-      // Delete server resources to prevent them from being exposed in the binary.
-      files.delete(key);
-    }
-  });
-
-  // TODO: Deprecate this in favor of a built-in prop that users should avoid setting.
-  const userDefinedServerUrl = exp.extra?.router?.origin;
-  let serverUrl = userDefinedServerUrl;
-
-  const shouldSkipServerDeployment = (() => {
-    if (!options.eager) {
-      logInXcode('Skipping server deployment because the script is not running in eager mode.');
-      return true;
-    }
-
-    // Add an opaque flag to disable server deployment.
-    if (env.EXPO_NO_DEPLOY) {
-      warnInXcode('Skipping server deployment because environment variable EXPO_NO_DEPLOY is set.');
-      return true;
-    }
-
-    // Can't safely deploy from Xcode since the PATH isn't set up correctly. We could amend this in the future and allow users who customize the PATH to deploy from Xcode.
-    if (isExecutingFromXcodebuild()) {
-      // TODO: Don't warn when the eager bundle has been run.
-      warnInXcode(
-        'Skipping server deployment because the build is running from an Xcode run script. Build with Expo CLI or EAS Build to deploy the server automatically.'
-      );
-      return true;
-    }
-
-    return false;
-  })();
-
-  // Deploy the server output to a hosting provider.
-  const deployedServerUrl = shouldSkipServerDeployment
-    ? false
-    : await runServerDeployCommandAsync(projectRoot, {
-        distDirectory: serverOutput,
-        deployScript: getServerDeploymentScript(pkg.scripts, options.platform),
-      });
-
-  if (!deployedServerUrl) {
-    Log.log(
-      `Skipping writing generated server URL (${deployedServerUrl}) to app.json because a value is already defined.`
-    );
-    return;
-  }
-
-  if (serverUrl) {
-    logInXcode(
-      `Using custom server URL: ${serverUrl} (ignoring deployment URL: ${deployedServerUrl})`
-    );
-  }
-
-  // If the user-defined server URL is not defined, use the deployed server URL.
-  // This allows for overwriting the server URL in the project's native files.
-  serverUrl ||= deployedServerUrl;
-
-  // If the user hasn't manually defined the server URL, write the deployed server URL to the app.json.
-  if (userDefinedServerUrl) {
-    return;
-  }
-  Log.log('Writing generated server URL to app.json');
-
-  // NOTE: Is is it possible to assert that the config needs to be modifiable before building the app?
-  const modification = await modifyConfigAsync(
-    projectRoot,
-    {
-      extra: {
-        ...(exp.extra ?? {}),
-        router: {
-          ...(exp.extra?.router ?? {}),
-          generatedOrigin: serverUrl,
-        },
-      },
-    },
-    {
-      skipSDKVersionRequirement: true,
-    }
-  );
-
-  if (modification.type !== 'success') {
-    throw new CommandError(
-      `Failed to write generated server origin to app.json because the file is dynamic and does not extend the static config. The client will not be able to make server requests to API routes or static files. You can disable server linking with EXPO_NO_DEPLOY=1 or by disabling server output in the app.json.`
-    );
-  }
-}
-
-/** We can try to remove the generated origin from the manifest when running outside of eager mode. Bundling is the last operation to run so the config will already be embedded with the origin. */
-async function tryRemovingGeneratedOriginAsync(projectRoot: string, exp: ExpoConfig) {
-  if (env.CI) {
-    // Skip in CI since nothing is committed.
-    return;
-  }
-  if (exp.extra?.router?.generatedOrigin == null) {
-    debug('No generated origin needs removing');
-    return;
-  }
-
-  const modification = await modifyConfigAsync(
-    projectRoot,
-    {
-      extra: {
-        ...(exp.extra ?? {}),
-        router: {
-          ...(exp.extra?.router ?? {}),
-          generatedOrigin: undefined,
-        },
-      },
-    },
-    {
-      skipSDKVersionRequirement: true,
-    }
-  );
-
-  if (modification.type !== 'success') {
-    debug('Could not remove generated origin from manifest');
-  } else {
-    debug('Generated origin has been removed from manifest');
-  }
 }
 
 export async function exportEmbedBundleAndAssetsAsync(
