+++ packages/@expo/cli/src/utils/env.ts
@@ -94,11 +94,6 @@ class Env {
     return string('EXPO_PUBLIC_FOLDER', 'public');
   }
 
-  /** Disable server deployment during production builds (during `expo export:embed`). This is useful for testing API routes and server components against a local server. */
-  get EXPO_NO_DEPLOY(): boolean {
-    return boolish('EXPO_NO_DEPLOY', false);
-  }
-
   /** Higher priority `$EDIOTR` variable for indicating which editor to use when pressing `o` in the Terminal UI. */
   get EXPO_EDITOR(): string {
     return string('EXPO_EDITOR', '');
@@ -218,6 +213,11 @@ class Env {
   /** Internal key used to pass eager bundle data from the CLI to the native run scripts during `npx expo run` commands. */
   get __EXPO_EAGER_BUNDLE_OPTIONS() {
     return string('__EXPO_EAGER_BUNDLE_OPTIONS', '');
+  }
+
+  /** Disable server deployment during production builds (during `expo export:embed`). This is useful for testing API routes and server components against a local server. */
+  get EXPO_NO_DEPLOY(): boolean {
+    return boolish('EXPO_NO_DEPLOY', false);
   }
 }
